chore: Rewrite "Add/Remove In Playlist" function
- We now have a toggle function and a single query instead of 2 since both queries are always used together.
- I think we didn't go with this method before due to not wanting a toggle method (in addition to having a separate query for toggling the favorites status).
